==> code/1_dataset/1_data_retrival.ipynb <==
from pybliometrics.scopus import AbstractRetrieval, config
from pybliometrics.scopus.exception import Scopus429Error
from pybliometrics.scopus.utils import create_config
import pandas as pd
from itertools import chain
from functools import reduce
import logging
import json
import glob
from tqdm import tqdm

create_config()

_keys = ['your_api_key']

logger = logging.getLogger()
logger.setLevel(logging.INFO)

doi.doi.values

with open('../../data/doi.csv', 'r') as file:
    doi = file.readlines()

def get_article_info(identifier):
    try:
        article_info = AbstractRetrieval(identifier=identifier,view='FULL')
        dict_info = {'title': article_info.title,
                     'abstract': article_info.abstract,
                     'affiliation': article_info.affiliation,
                     'aggregationType': article_info.aggregationType,
                     'authkeywords': article_info.authkeywords,
                     'authorgroup': article_info.authorgroup,
                     'authors': article_info.authors,
                     'chemicals': article_info.chemicals,
                     'citedby_count': article_info.citedby_count,
                     'coverDate': article_info.coverDate,
                     'description': article_info.description,
                     'doi': article_info.doi,
                     'funding': article_info.funding,
                     'funding_text': article_info.funding_text,
                     'citedby_link': article_info.citedby_link,
                     'contributor_group': article_info.contributor_group,
                     'language': article_info.language,
                     'publicationName': article_info.publicationName,
                     'references': article_info.references,
                     'subject_areas': article_info.subject_areas,
                     'issueIdentifier': article_info.issueIdentifier,
                     'volume': article_info.volume,
                     'eid': article_info.eid}
        logger.info('Info from: {} recived'.format(identifier))
    except Scopus429Error:
        # Remove the last item in _keys to assign it as new API key
        config["Authentication"]["APIKey"] = _keys.pop()
        logger.warning('changed API key')
        article_info = AbstractRetrieval(identifier=identifier,view='FULL')
        dict_info = {'title': article_info.title,
                     'abstract': article_info.abstract,
                     'affiliation': article_info.affiliation,
                     'aggregationType': article_info.aggregationType,
                     'authkeywords': article_info.authkeywords,
                     'authorgroup': article_info.authorgroup,
                     'authors': article_info.authors,
                     'chemicals': article_info.chemicals,
                     'citedby_count': article_info.citedby_count,
                     'coverDate': article_info.coverDate,
                     'description': article_info.description,
                     'doi': article_info.doi,
                     'funding': article_info.funding,
                     'funding_text': article_info.funding_text,
                     'citedby_link': article_info.citedby_link,
                     'contributor_group': article_info.contributor_group,
                     'language': article_info.language,
                     'publicationName': article_info.publicationName,
                     'references': article_info.references,
                     'subject_areas': article_info.subject_areas,
                     'issueIdentifier': article_info.issueIdentifier,
                     'volume': article_info.volume,
                     'eid': article_info.eid}
        logger.info('Info from: {} recived'.format(identifier))
    except Exception as e:
        logger.error('Failed at {}'.format(identifier))
        logger.error('Original message: {}'.format(e))
        dict_info = dict()
    finally:
        return dict_info

example = get_article_info('10.1177/0162243915596231')

example

example= pd.DataFrame.from_dict(example,orient='index').transpose()

example.affiliation

2 paper's example

articles = list(map(lambda x: get_article_info(x),doi.values))

articles = list(filter(None, articles))

df = pd.DataFrame(df)

df.to_picke('../../data/intermediate_files/dataset.p')

---------
