==> code/1_dataset/2_data_cleaning.ipynb <==
import pandas as pd
from itertools import chain
from functools import reduce
import numpy as np
import logging
import json
import glob
from tqdm import tqdm
import re

df = pd.read_pickle('../../data/intermediate_files/dataset.p')

df.columns

def clean_publicationName(text):
    text = text.lower()
    text = re.sub('(?<=studies in history and philosophy of science).*','',text)
    text = re.sub('(?<=isis).*','',text)
    text = re.sub('(?<=synthese).*','',text)
    text = re.sub('(?<=scientometrics).*','',text)
    text = re.sub('.*(?=british journal for the history of science)','',text)
    text = re.sub('science(,)? technology (&|and) society','science, technology and society',text)
    text = re.sub('science(,)? technology(,)? (&|and) human values','science, technology and human values',text)
    text = re.sub('science & education','science and education',text)
    return text
    

df.publicationName = df.publicationName.apply(lambda x: clean_publicationName(x))

df.publicationName.value_counts()

def normalize_publicationName(argument): 
    switcher = { 
        'Science, Technology & Human Values': "Science Technology and Human Values", 
        'Science, Technology, & Human Values': "Science Technology and Human Values", 
        'Science, technology & human values': "Science Technology and Human Values",
        'Scientometrics: An International Journal for all Quantitative Aspects of the Science of Science, Communication in Science and Science Policy': "Scientometrics",
        'Computers and Education' : 'Research Policy',
        'Public understanding of science (Bristol, England)':'Public Understanding of Science',
        
        'Social studies of science': 'Social Studies of Science'
    } 
    return switcher.get(argument,argument) 

df.publicationName = df.publicationName.apply(normalize_publicationName)

filter articles incorrectly downloaded

df = df[~df['publicationName'].isin(['lecture notes in computer science (including subseries lecture notes in artificial intelligence and lecture notes in bioinformatics)',
                                     'Automation in Construction',
                              'science advances']) ]


#### Distribution of citatiosn

def most_cited(series):

    return set(df.title[df.citedby_count == max(series)].values)

df.groupby(['publicationName']).citedby_count.agg(['count', 'mean',np.sum, 'max', most_cited]).sort_values(by ='mean')

### affiliation

affiliation = list(filter(None, df['affiliation'])) 
affiliation = [item for sublist in affiliation for item in sublist]

affiliation_df = pd.DataFrame(affiliation, columns=['id', 'name','city', 'country'])
affiliation_df.drop_duplicates(['id', 'name','city', 'country'],inplace = True)
affiliation_df.to_json('../../data/affiliations.json')
affiliation_df.sample()

#### authors

authors = list(filter(None, df['authors'])) 
authors = [item for sublist in authors for item in sublist]
authors
authors_df = pd.DataFrame(authors, columns=['auid', 'indexed_name','surname', 'given_name','affiliation_id'])
authors_df.drop_duplicates(['auid', 'indexed_name','surname', 'given_name'],inplace = True)
authors_df.to_json('../../data/authors.json')
authors_df.sample()


#### references

def add_eid(ref_list, eid):
    if ref_list is not None:
        #ref_list = ref_list.tolist()[0]
        ref_list_extended = list(map(lambda x: x + [eid],ref_list))
    else:
        ref_list_extended = None
    return ref_list_extended

df["references"] = df.apply(lambda x: add_eid(x.references,x.eid), axis=1)

columns = ['position', 'eid_of_ref', 'doi', 'title', 'authors', 'authors_auid', 'authors_affiliationid', 'sourcetitle', 'publicationyear', 'volume', 'issue', 'first', 'last', 'citedbycount', 'type', 'text', 'fulltext','eid']

references = [x for x in df.references.values if x]
references = list(chain.from_iterable(references))

references_df = pd.DataFrame(references, columns=columns)

references_df.sample()

references_df.to_json('../../data/references.json')

df.to_pickle('../../data/intermediate_files/dataset.p')
