==> code/1_dataset/4_nlp_features_Bert.ipynb <==
#pip install transformers

import torch
from transformers import BertTokenizer, BertModel

# OPTIONAL: if you want to have more information on what's happening, activate the logger as follows
import logging
#logging.basicConfig(level=logging.INFO)

import matplotlib.pyplot as plt
import pandas as pd
from collections import Counter

import pandas as pd
import numpy as np
from sklearn.manifold import TSNE
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt
import seaborn as sns; sns.set(color_codes=True)
import os

# Load pre-trained model tokenizer (vocabulary)
tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

df = pd.read_pickle('../../data/dataset.p')

texts = df.text_clean
#add markers
texts = ["[CLS] " + text.lower() + " [SEP]" for text in texts]
#remove NUM
texts = [text.replace('num','') for text in texts]

#tokenize
tokenized_text = [tokenizer.tokenize(text) for text in texts]

def cut_text(tokens):
    if len(tokens)>=500:
        tokens = tokens[0:499] + ['[SEP]']
    return tokens

tokenized_text = list(map(cut_text,tokenized_text))

indexed_tokens = [tokenizer.convert_tokens_to_ids(text) for text in tokenized_text]

segments_ids = []
for i in range(len(indexed_tokens)):
    segments_id = [i] * len(indexed_tokens[i]) # maybe change [i] for i
    segments_ids.append(segments_id)



len(segments_ids)

#to tensor
tokens_tensor = [torch.tensor([t]) for t in indexed_tokens]
segments_tensors = [torch.tensor([t]) for t in segments_ids]

#load pretrained model
# Load pre-trained model (weights)
model = BertModel.from_pretrained('bert-base-uncased',
                                  output_hidden_states = True, # Whether the model returns all hidden-states.
                                )

# Put the model in "evaluation" mode, meaning feed-forward operation.
model.eval()


def sentence_embedding(tokens_tensor,segments_tensors):
    #Run the text through BERT, and collect all of the hidden states produced
    # from all 12 layers. 
    with torch.no_grad():

        outputs = model(tokens_tensor, segments_tensors)

        # Evaluating the model will return a different number of objects based on 
        # how it's  configured in the `from_pretrained` call earlier. In this case, 
        # becase we set `output_hidden_states = True`, the third item will be the 
        # hidden states from all layers. See the documentation for more details:
        # https://huggingface.co/transformers/model_doc/bert.html#bertmodel
        hidden_states = outputs[2]
        token_vecs = hidden_states[-2][0]
        sentence_embedding = torch.mean(token_vecs, dim=0)
        return sentence_embedding

sentences_embeddings = [sentence_embedding(x,y) for x,y in zip(tokens_tensor, segments_tensors)]

semb = sentences_embeddings.copy()

semb = [s.detach().numpy().tolist() for s in sentences_embeddings]

df['bert_embedding'] = semb


df.to_pickle('../../data/dataset.p')
