==> code/1_dataset/5_pyg_network_setup.ipynb <==
import pandas as pd
import pickle

from libs.NetworkBuilder import NetworkBuilder
from libs.ScisciDataset import ScisciDataset

df = pd.read_pickle('../../data/dataset.p')

ref_df = pd.read_json('../../data/references.json')

def save_dataset(dataset, file_name):
    with open(file_name, 'wb') as handle:
        pickle.dump(dataset, handle, protocol=pickle.HIGHEST_PROTOCOL)


# TF-IDF

nb = NetworkBuilder(df, references=ref_df)

X,LE = nb.prepare_metadata(text_encoding = 'tfidf')

edge_index,eid_row_indices = nb.build_edges()

save_dataset(edge_index, '../../data/torch/tfidf/edge_pairs.p')
save_dataset(X, '../../data/torch/tfidf/x.p')
save_dataset(eid_row_indices, '../../data/torch/tfidf/eid_row_indices.p')
save_dataset(LE, '../../data/torch/tfidf/LE.p')

x_tfidf = pd.read_pickle('../../data/torch/tfidf/x.p')

## D2V

df = pd.read_pickle('../../data/dataset.p')

nb = NetworkBuilder(df, references=ref_df)

X,LE = nb.prepare_metadata(text_encoding = 'd2v')

edge_index,eid_row_indices = nb.build_edges()

def filter_references(corpus_eid, references):
    #remove references without ID
    ref_df = references.copy()
    ref_df = ref_df[ref_df.eid_of_ref.notnull()].reset_index(drop=True)
    # add the '2-s2.0-' of the eid_of_ref
    ref_df['eid_of_ref'] = ref_df['eid_of_ref'].apply(lambda x: '2-s2.0-' + str(int(x)))
    #filter for within corpus eids
    ref_df = ref_df[ref_df['eid_of_ref'].isin(corpus_eid)]
    ref_df = ref_df[ref_df['eid'].isin(corpus_eid)]
    return ref_df

filter_ref_df = filter_references(df.eid.unique(),ref_df)

network_ids = set(filter_ref_df.eid_of_ref).union(set(filter_ref_df.eid))

with open('../../results/network_ids.txt', 'w') as f:
    for item in list(network_ids):
        f.write("%s\n" % item)

save_dataset(edge_index, '../../data/torch/d2v/edge_pairs.p')
save_dataset(X, '../../data/torch/d2v/x.p')
save_dataset(eid_row_indices, '../../data/torch/d2v/eid_row_indices.p')
save_dataset(LE, '../../data/torch/d2v/LE.p')

x_d2v = pd.read_pickle('../../data/torch/d2v/x.p')

x_d2v.shape

# dataset
dataset = ScisciDataset(root='test',raw_path = '../../data/torch/d2v')
data = dataset[0]
#data = train_test_split_edges(data)

data

# BERT

nb = NetworkBuilder(df, references=ref_df)

X,LE = nb.prepare_metadata(text_encoding = 'bert')

edge_index,eid_row_indices = nb.build_edges()

X.shape

save_dataset(edge_index, '../../data/torch/bert/edge_pairs.p')
save_dataset(X, '../../data/torch/bert/x.p')
save_dataset(eid_row_indices, '../../data/torch/bert/eid_row_indices.p')
save_dataset(LE, '../../data/torch/bert/LE.p')
