==> code/1_dataset/6_data_statistics.ipynb <==
import pandas as pd
import numpy as np
import pickle
from igraph import *
from tqdm.notebook import tqdm
import matplotlib
%matplotlib inline



df = pd.read_pickle('../../data/dataset.p')

### 1.Subject areas

journals = df.filter(['publicationName','subject_areas'])

journals['subject_areas'] = journals.apply(lambda x: [l[0] for l in x.subject_areas],axis=1)

journals[['subject_areas_0','subject_areas_1','subject_areas_2','subject_areas_3','subject_areas_4','subject_areas_5']] = pd.DataFrame(journals.subject_areas.tolist())

journals.subject_areas = None

## subfields

quant = ['scientometrics','journal of informetrics']
hist_phil = ['british journal for the history of science','isis','science and education',            
             'science and technology studies','science, technology and society','social studies of science',
             'studies in history and philosophy of science','synthese']
management = ['research policy','science and public policy']
social_sci = ['minerva','public understanding of science','research evaluation', 'science, technology and human values']

df.loc[df.publicationName.isin(quant), 'field'] = 'Library and Information Sciences'
df.loc[df.publicationName.isin(management), 'field'] = 'Management'
df.loc[df.publicationName.isin(social_sci), 'field'] = 'Education, Communication and Anthropology'
df.loc[df.publicationName.isin(hist_phil), 'field'] = 'History and Philosophy'
#df.loc[df.publicationName.isin(education), 'field'] = 'Education'



def most_cited(series):

    return set(df.title[df.citedby_count == max(series)].values)

def mean_round(x):
    return round(np.mean(x),2)

def make_summary(df):

    summary_stats = df.groupby(['publicationName','field']).citedby_count.agg(['count', mean_round, 'max']).sort_values(by =['field','mean_round'])
    summary_stats.reset_index(inplace=True)

    summary_stats['publicationName'] = summary_stats.apply(lambda x: x.publicationName.title(),axis=1)
    total = ['Total',' ',summary_stats['count'].sum(),mean_round(summary_stats['mean_round']),summary_stats['max'].max()]

    summary_stats.loc['Total']= total

    summary_stats = summary_stats.sort_values(by=['field','count'], ascending=False)

    summary_stats.set_index(['field', 'publicationName'],inplace=True)
    return summary_stats

summary_stats = make_summary(df)

print(summary_stats.to_latex(index=True))

# 2.Network

raw_dir = '../../data/torch/bert/'
names = ['x.p', 'edge_pairs.p', 'eid_row_indices.p']
x,edge_pairs,eid_row_indices = [pickle.load(open('{}/{}'.format(raw_dir, name), 'rb'))for name in names]



edge_pairs = edge_pairs.detach().numpy()

edge_pairs_df = pd.DataFrame(edge_pairs).T

_ = edge_pairs_df.merge(eid_row_indices, left_on=0, right_on='index')
_ = _.merge(eid_row_indices, left_on=1, right_on='index')

edges = _.loc[:,['eid_x','eid_y']]

g = Graph.TupleList(edges.values,directed=False)

descriptive statistics

g.average_path_length()

np.mean(g.degree())

np.max(g.degree())

g.vcount()

len(set(edges.eid_x).union(set(edges.eid_y)))

g.ecount()

g.assortativity_degree()

np.mean(g.betweenness())

g.is_connected()

comps= g.components()

giant = comps.giant()

giant.ecount()

giant.vcount()

print('Edges not in the giant component: {}'.format(g.ecount()-giant.ecount()))
print('Vertices not in the giant component: {}'.format(g.vcount()-giant.vcount()))

np.mean(giant.closeness())

np.mean(giant.eigenvector_centrality())
np.mean(g.eigenvector_centrality())

g.farthest_points()

g.diameter()

g.summary()

### small word

#clustering coefficient
clust_coef = g.transitivity_undirected()
clust_coef

average_path_length = g.average_path_length()
average_path_length

### random model

def random_model(n = g.vcount(), m = g.ecount(), times=100):
    rnd_clust_coefs = []
    rnd_lengths = []
    for i in tqdm(range(times)):
        rnd_graph = Graph.Erdos_Renyi(n=n,m=m)
        clust = rnd_graph.transitivity_undirected()
        length = rnd_graph.average_path_length()
        rnd_clust_coefs += [clust]
        rnd_lengths += [length]
    return rnd_clust_coefs, rnd_lengths

rnd_clust_coefs, rnd_lengths = random_model()

mean_random_clust_coef = np.mean(rnd_clust_coefs)
mean_random_clust_coef

mean_random_length = np.mean(rnd_lengths)
mean_random_length

clust_coef/mean_random_clust_coef

average_path_length / mean_random_length

average_path_length

### 3. Differences between dataset and network 

network_ids = [line.rstrip('\n') for line in open('../../results/network_ids.txt')]

in_net_df = df[df.eid.isin(network_ids)]
out_net_df = df[-df.eid.isin(network_ids)]

out_net_df.citedby_count.mean()

out_net_df.loc[:,'coverDate'] = out_net_df.loc[:,'coverDate'].apply(lambda x: int(x.rsplit('-')[0]))

out_net_df.coverDate.mean()

#### Out of network stats

oon_summary_stats = make_summary(out_net_df)
oon_summary_stats

#### In network stats

in_net_summary = make_summary(in_net_df)
in_net_summary

# Topic Modelling distribution

LDA = LDA_wrapper()

lda_model = LDA.restore_model("../../results/lda_model20t.p")
texts = df['text_clean'].values
data_vectorized, vectorizer = LDA.data_vectorizer(data=texts)

def print_topics(model, count_vectorizer, n_top_words):
    words = count_vectorizer.get_feature_names()
    for topic_idx, topic in enumerate(model.components_):
        print("\nTopic #%d:" % topic_idx)
        print(" ".join([words[i]
                        for i in topic.argsort()[:-n_top_words - 1:-1]]))

print_topics(lda_model, vectorizer, 5)


selected topics: 
The LDAViz change the topics order, so we need to map them. 


|LDAViz | lda_model|
|---|--|
|1 | 18|
|2 | 6|
|3 | 16|
|4 | 11|
|5 | 19|
|6 | 7|
|7 | 16|
|8 | 9|
|9 | 8|
|10 | 14|
|11 | 10|
|12 | 12|
|13 | 5|
|14 | 13|
|15 | 2|
|16 | 4|
|17 | 3|
|18 | 0|
|19 | 15|
|20 | 1|



def get_topics(x):
    topic_1 = x[18]
    topic_2 = x[6]
    topic_3 = x[16]
    topic_4 = x[11]
    topic_5 = x[19]
    topic_6 = x[7]
    topic_7 = x[16]
    topic_8 = x[9]
    topic_9 = x[8]
    topic_10 = x[14]
    topic_11 = x[10]
    topic_12 = x[12]
    topic_13 = x[5]
    topic_14 = x[13]
    topic_15 = x[2]
    topic_16 = x[4]
    topic_17 = x[3]
    topic_18 = x[0]
    topic_19 = x[15]
    topic_20 = x[1]


    return{'topic_1':topic_1,
    'topic_2':topic_2,
    'topic_3':topic_3,
    'topic_4':topic_4,
    'topic_5':topic_5,
    'topic_6':topic_6,
    'topic_7':topic_7,
    'topic_8':topic_8,
    'topic_9':topic_9,
    'topic_10':topic_10,
    'topic_11':topic_11,
    'topic_12':topic_12,
    'topic_13':topic_13,
    'topic_14':topic_14,
    'topic_15':topic_15,
    'topic_16':topic_16,
    'topic_17':topic_17,
    'topic_18':topic_18,
    'topic_19':topic_19,
    'topic_20':topic_20}
    

topics_df = pd.DataFrame(list(df.apply(lambda x: get_topics(x.topic_dist), axis=1)))

topics_df[['publicationName','field']] = df[['publicationName','field']].reset_index(drop=True)

topics_summary = topics_df.groupby(['publicationName','field']).agg(mean)
topics_df.to_csv('../../results/topics_dist_by_article.csv', index=False)
